MainMemory - SAModel: New definitions for Cube, Dash Panel, Springs. Ring Group can't be selected with the mouse for some reason.

+++ SADXLVL/DataTypes.cs
@@ -220,9 +220,9 @@ namespace SonicRetro.SAModel.SADXLVL2
             bytes.AddRange(BitConverter.GetBytes(ID));
             unchecked
             {
-                bytes.AddRange(BitConverter.GetBytes(Rotation.X));
-                bytes.AddRange(BitConverter.GetBytes(Rotation.Y));
-                bytes.AddRange(BitConverter.GetBytes(Rotation.Z));
+                bytes.AddRange(BitConverter.GetBytes((ushort)Rotation.X));
+                bytes.AddRange(BitConverter.GetBytes((ushort)Rotation.Y));
+                bytes.AddRange(BitConverter.GetBytes((ushort)Rotation.Z));
             }
             bytes.AddRange(Position.GetBytes());
             bytes.AddRange(Scale.GetBytes());
